Export define on app ready
